+++ src/lips.js
@@ -22,7 +22,8 @@
 })(typeof window !== 'undefined' ? window : global, function(root, BN, undefined) {
     // parse_argument based on function from jQuery Terminal
     var re_re = /^\/((?:\\\/|[^/]|\[[^\]]*\/[^\]]*\])+)\/([gimy]*)$/;
-    var float_re = /^[-+]?[0-9]*\.?[0-9]*([eE][-+]?[0-9]+)?$/;
+    var int_re = /^[-+]?[0-9]+([eE][-+]?[0-9]+)?$/;
+    var float_re = /^([-+]?((\.[0-9]+|[0-9]+\.[0-9]+)([eE][-+]?[0-9]+)?)|[0-9]+\.)$/;
     // ----------------------------------------------------------------------
     function parse_argument(arg) {
         function parse_string(string) {
@@ -45,8 +46,8 @@
             return new RegExp(regex[1], regex[2]);
         } else if (arg.match(/['"]/)) {
             return parse_string(arg);
-        } else if (arg.match(/^-?[0-9]+$/)) {
-            return LNumber(parseInt(arg, 10));
+        } else if (arg.match(int_re)) {
+            return LNumber(parseFloat(arg));
         } else if (arg.match(float_re)) {
             return LNumber(parseFloat(arg), true);
         } else if (arg === 'nil') {
@@ -556,7 +557,7 @@
         }
         if (typeof this !== 'undefined' && this.constructor !== LNumber ||
             typeof this === 'undefined') {
-            return new LNumber(n, float);
+            return new LNumber(n, float === true ? true : undefined);
         }
         if (!LNumber.isNumber(n)) {
             throw new Error("You can't create LNumber from " + typeof n);
@@ -1900,6 +1901,19 @@
     } else if (typeof window !== 'undefined') {
         global_env.set('window', window);
     }
+    function type(value) {
+        if (typeof value === "string") {
+            return "string";
+        } else if (value instanceof LNumber) {
+            return "number";
+        } else if (value instanceof RegExp) {
+            return "regex";
+        } else if (typeof value === 'boolean') {
+            return 'boolean';
+        } else {
+            return 'unknown type';
+        }
+    }
     // ----------------------------------------------------------------------
     function evaluate(code, {env, dynamic_scope, error = () => {}} = {}) {
         try {
@@ -1977,6 +1991,9 @@
                     throw new Error('Unbound variable `' + code.name + '\'');
                 }
                 return value;
+            } else if (code instanceof Pair) {
+                value = first.toString();
+                throw new Error(`${type(first)} ${value} is not a function`);
             } else {
                 return code;
             }