+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * fix defmacro
 * fix parsing expressions that have multiple special characters in a row
 * handle float calculations
+* throw exception when try to execute native non function value
 
 ## 0.5.4
 ### Bugs